package extractor

+++ helpers/helpers_emba_print.sh
@@ -595,8 +595,10 @@ module_end_log() {
 
   # check if there is some content in the csv log file. If there is only
   # one entry line we remove the file at all
-  if [[ $(wc -l "$CSV_LOG" | awk '{print $1}') -lt 2 ]]; then
-    rm "$CSV_LOG"
+  if [[ -f "$CSV_LOG" ]]; then
+    if [[ $(wc -l "$CSV_LOG" | awk '{print $1}') -lt 2 ]]; then
+      rm "$CSV_LOG"
+    fi
   fi
 
   print_output "[*] $(date) - $MODULE_MAIN_NAME finished" "main"
